perf: added server fetch to title data
